// src/types/report.d.ts
export interface NewHazardReport {
  // Report Data
  reportTitle: string;
  coordinates: string;
  reportDescription: string;
  severity: "LOW" | "MEDIUM" | "HIGH"; // assuming possible values
  mainType: "HAZARD";
  status: "INITIATED" | "IN_PROGRESS" | "RESOLVED" | string;

  // Incident Data
  categoryType: string;
  hazardDescription: string;

  managerSignatureConfirmationDate: string | null;
  dynamicQuestion: {
    questionId: string;
    answer: string;
  }[];
  // Media (UUIDs or URLs)
  media: string[];
}
// ...
export enum severity {
  LOW = "LOW",
  MEDIUM = "MEDIUM",
  HIGH = "HIGH",
  EXTREME = "EXTREME",
}